requirements.txt
# ========== Flask Core ==========
Flask==3.0.0
Flask-Cors==4.0.0
Werkzeug==3.1.3
Jinja2==3.1.6
click==8.3.1
itsdangerous==2.2.0
MarkupSafe==3.0.3
blinker==1.9.0

# ========== Security & Crypto ==========
bcrypt==5.0.0
cryptography==45.0.5
pycryptodome==3.23.0
pyOpenSSL==25.1.0
cffi==1.17.1
pycparser==2.22

# ========== HTTP & Web ==========
requests==2.32.4
urllib3==2.4.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
beautifulsoup4==4.13.4
soupsieve==2.7

# ========== Network Scanning ==========
python-nmap==0.7.1
netifaces==0.11.0

# ========== Git Analysis ==========
GitPython==3.1.43

# ========== Environment & Config ==========
python-dotenv==1.2.1

# ========== Server (Production) ==========
gunicorn==21.2.0

# ========== Data Processing ==========
numpy==2.3.0
pandas==2.3.0

# ========== Visualization (optionnel) ==========
matplotlib==3.10.3
pillow==11.2.1

# ========== Date & Time ==========
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# ========== System Monitoring ==========
psutil==5.9.8

# ========== Utilities ==========
attrs==25.3.0
packaging==25.0
six==1.17.0
typing_extensions==4.14.0

# ========== Dependencies ==========
aioquic==1.2.0
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.2
kiwisolver==1.4.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pylsqpack==0.3.22
pyparsing==3.2.3
service-identity==24.2.0